requirements.txt
openai>=1.3.0
python-dotenv>=1.0.0
selenium>=4.15.0
beautifulsoup4>=4.12.0
requests>=2.31.0
imaplib2>=3.10
webdriver-manager>=4.0.0
